Use same design for all player select buttons
