+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/home/players/PlayerSelectionButton.kt
@@ -1,67 +1,92 @@
 package io.music_assistant.client.ui.compose.home.players
 
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.defaultMinSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Speaker
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Button
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import io.music_assistant.client.data.model.client.PlayerData
-import io.music_assistant.client.data.model.client.PlayerDataFixtures
+import io.music_assistant.client.player.sendspin.SendspinState
+import io.music_assistant.client.ui.compose.home.HomeScreenViewModel
 
 @Composable
 fun PlayerSelectionButton(
     player: PlayerData,
+    playersState: HomeScreenViewModel.PlayersState.Data,
     onSelectPlayer: () -> Unit = {}
 ) {
-    Column(
-        modifier = Modifier
-            .clickable(onClick = onSelectPlayer)
-            .defaultMinSize(
-                minWidth = ButtonDefaults.MinWidth,
-                minHeight = ButtonDefaults.MinHeight
-            ).padding(8.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Icon(
-            modifier = Modifier.padding(top = 8.dp),
-            imageVector = Icons.Default.Speaker,
-            tint = MaterialTheme.colorScheme.primary,
-            contentDescription = null
-        )
+    val isLocalPlayer = player.playerId == playersState.localPlayerId
+    val dotColor = (if (isLocalPlayer) playersState.sendspinState else null)
+        .toDotColor()
+        .takeIf { isLocalPlayer }
 
-        Text(
-            player.player.displayName,
-            modifier = Modifier.padding(top = 4.dp),
-            color = MaterialTheme.colorScheme.primary,
-            style = MaterialTheme.typography.labelSmall
-        )
+    Box {
+        val playerName: @Composable (Color) -> Unit = { textColor ->
+            Row(
+                modifier = Modifier.align(Alignment.Center),
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                Text(
+                    text = player.player.displayName + (if (isLocalPlayer) " (local)" else ""),
+                    style = MaterialTheme.typography.bodyMedium,
+                    textAlign = TextAlign.Center,
+                    fontWeight = FontWeight.Medium,
+                    color = textColor,
+                    maxLines = 1,
+                    overflow = TextOverflow.Ellipsis
+                )
+                dotColor?.let {
+                    Spacer(modifier = Modifier.width(4.dp))
+                    Box(
+                        modifier = Modifier
+                            .size(8.dp)
+                            .background(it, CircleShape)
+                    )
+                }
+            }
+        }
+
+        when {
+            player.groupChildren.none { it.isBound } ->
+                OutlinedButton(
+                    modifier = Modifier.align(Alignment.Center),
+                    enabled = true,
+                    onClick = onSelectPlayer
+                ) {
+                    playerName(MaterialTheme.colorScheme.onSurface)
+                }
+
+            else ->
+                Button(
+                    modifier = Modifier.align(Alignment.Center),
+                    enabled = true,
+                    onClick = onSelectPlayer
+                ) {
+                    playerName(MaterialTheme.colorScheme.onPrimary)
+                }
+        }
     }
 }
 
-@Preview
-@Composable
-fun PreviewPlayerSelectionButton() {
-    val player = PlayerDataFixtures.playerData()
-    PlayerSelectionButton(
-        player = player
-    )
-}
-
-@Preview
-@Composable
-fun PreviewPlayerSelectionButtonLongName() {
-    val player = PlayerDataFixtures.playerData(name = "Very Long Speaker Name")
-    PlayerSelectionButton(
-        player = player
-    )
+private fun SendspinState?.toDotColor(): Color = when (this) {
+    is SendspinState.Synchronized, is SendspinState.Ready,
+    is SendspinState.Buffering -> Color(0xFF4CAF50) // Green
+    is SendspinState.Connecting, is SendspinState.Authenticating,
+    is SendspinState.Handshaking, is SendspinState.Reconnecting -> Color(0xFFFF9800) // Orange
+    is SendspinState.Error -> Color(0xFFF44336) // Red
+    is SendspinState.Idle, null -> Color(0xFFBDBDBD) // Light gray
 }
